requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.0
anthropic==0.3.6
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
azure-core==1.29.6
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
boto3==1.21.22
botocore==1.24.46
cachetools==5.3.2
camel-converter==3.0.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7
cloudpickle==2.2.1
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
databricks-cli==0.17.7
databricks-vectorsearch==0.22
dataclasses-json==0.5.14
datasets==2.15.0
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
Django==4.2.4
dnspython==2.4.2
docker==6.1.3
email-validator==1.3.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.105.0
fastjsonschema==2.18.0
filelock==3.12.4
fire==0.4.0
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.34
google-api-core==2.15.0
google-api-python-client==2.110.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
googleapis-common-protos==1.62.0
greenlet==2.0.2
grpcio==1.60.0
grpcio-tools==1.57.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-server==1.24.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab==3.3.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.344
langchain-core==0.0.13
langchainplus-sdk==0.0.20
langsmith==0.0.69
libcst==1.0.1
llama-index==0.9.15.post2
loguru==0.6.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
meilisearch==0.21.0
mistune==3.0.1
mlflow==2.9.1
mlflow-skinny==2.9.1
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==0.5.6
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.4.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
orjson==3.9.5
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.10.0
pluggy==1.2.0
portalocker==2.7.0
posthog==3.1.0
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.39
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pygame==2.5.1
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.0.2
pyparsing==3.1.1
pypdf==3.17.3
PyPika==0.48.9
pysbd==0.3.4
pyspark==3.5.0
pytest==7.2.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2023.3
pyvis==0.3.2
PyYAML==6.0
pyzmq==25.1.1
qdrant-client==1.4.0
querystring-parser==1.2.4
ragas==0.0.22
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.0
rsa==4.9
s3transfer==0.5.2
safetensors==0.4.0
scikit-learn==1.0.2
scipy==1.11.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.23
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.3.3
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
triton==2.1.0
typer==0.9.0
types-protobuf==4.24.0.4
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.18.3
uvloop==0.17.0
validators==0.22.0
watchfiles==0.20.0
wcwidth==0.2.6
weaviate-client==3.25.3
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.16.2
